+++ SWP291_Team2_SE1829NJ/src/java/controller/acc/accCreate.java
@@ -3,7 +3,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
 
-package controller.admin;
+package controller.acc;
 
 import dao.AccountDAO;
 import java.io.IOException;
@@ -14,7 +14,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import java.sql.Date;
-import model.Account;
 
 /**
  *
