+++ SWP291_Team2_SE1829NJ/src/java/controller/acc/accDelete.java
@@ -3,7 +3,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
 
-package controller.admin;
+package controller.acc;
 
 import dao.AccountDAO;
 import java.io.IOException;
@@ -13,56 +13,33 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import java.util.List;
 import model.Account;
 
 /**
  *
  * @author HUNG
  */
-public class accDetail extends HttpServlet {
-   
-    /** 
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
+public class accDelete extends HttpServlet {
+
+    /**
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
-     */protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-        throws ServletException, IOException {
-    response.setCharacterEncoding("UTF-8");
-    response.setContentType("text/html; charset=UTF-8");
+     */
+    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+        response.setContentType("text/html;charset=UTF-8");
 
-    try {
-        HttpSession session = request.getSession(); 
-        model.Account acc = (Account) session.getAttribute("acc");
-        
-        // Kiểm tra vai trò của người dùng (ví dụ: admin là role_id = 1)
-        if (acc.getRole_id().equalsIgnoreCase("1")) {
-            // Lấy username từ request parameter
-            String username = request.getParameter("username");
-            
-            // Sử dụng AccountDAO để lấy thông tin chi tiết của tài khoản theo username
-            AccountDAO udao = new AccountDAO();
-            Account acc1 = udao.getAnAccountByUsername(username);
-
-            // Đặt thuộc tính vào request để truyền sang accountDetail.jsp
-            request.setAttribute("acc1", acc1);
-
-            // Chuyển hướng (forward) request đến trang accountDetail.jsp
-            request.getRequestDispatcher("/admin/accountDetail.jsp").forward(request, response);
-        } else {
-            // Nếu không phải admin, chuyển hướng đến trang login
-            response.sendRedirect("login");
-        }
-    } catch (ServletException | IOException e) {
-        response.sendRedirect("404.jsp");
     }
-}
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /** 
+    /**
      * Handles the HTTP <code>GET</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -70,12 +47,13 @@ public class accDetail extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
-    } 
+    }
 
-    /** 
+    /**
      * Handles the HTTP <code>POST</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -83,12 +61,38 @@ public class accDetail extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
-        processRequest(request, response);
+            throws ServletException, IOException {
+        response.setCharacterEncoding("UTF-8");
+        response.setContentType("text/html; charset=UTF-8");
+
+        try {
+            HttpSession session = request.getSession();
+            model.Account acc = (Account) session.getAttribute("acc");
+            if (acc.getRole_id().equalsIgnoreCase("1")) {
+
+//             
+                String namedele = request.getParameter("username");
+                System.out.println("Username to delete: " + namedele); // Kiểm tra giá trị nhận được
+                if (namedele != null) {
+                    AccountDAO udao = new AccountDAO();
+                    udao.AccDelete(namedele);
+                    request.getRequestDispatcher("accountmanagement").forward(request, response);
+                } else {
+                    System.out.println("No username received for deletion.");
+                    response.sendRedirect("errorPage.jsp"); // Chuyển hướng đến trang lỗi nếu không có dữ liệu
+                }
+
+            } else {
+                response.sendRedirect("login");
+            }
+        } catch (ServletException | IOException e) {
+            response.sendRedirect("404.jsp");
+        }
     }
 
-    /** 
+    /**
      * Returns a short description of the servlet.
+     *
      * @return a String containing servlet description
      */
     @Override
