+++ SWP291_Team2_SE1829NJ/src/java/controller/acc/accEdit.java
@@ -2,15 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-package controller.admin;
+
+package controller.acc;
 
 import dao.AccountDAO;
-import dao.BillDAO;
-import dao.BlogDAO;
-import dao.MenuDAO;
-import dao.RoomDAO;
-import dao.UserAccountDAO;
-import dao.UserDAO;
 import java.io.IOException;
 import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
@@ -18,22 +13,20 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import java.time.LocalDate;
-import java.util.List;
+import java.sql.Date;
+import java.sql.SQLException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import model.Account;
-import model.Booking;
-import model.UserAccount;
 
 /**
  *
  * @author HUNG
  */
-public class accountManagement extends HttpServlet {
-
-    /**
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
-     * methods.
-     *
+public class accEdit extends HttpServlet {
+   
+    /** 
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -41,63 +34,41 @@ public class accountManagement extends HttpServlet {
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-  
+        response.setContentType("text/html;charset=UTF-8");
+
         response.setCharacterEncoding("UTF-8");
         response.setContentType("text/html; charset=UTF-8");
 
         try {
-            HttpSession session = request.getSession(); 
+            HttpSession session = request.getSession();
             model.Account acc = (Account) session.getAttribute("acc");
             if (acc.getRole_id().equalsIgnoreCase("1")) {
 
 //             
+                String namedele = (String) request.getAttribute("usernamedele");
                 AccountDAO udao = new AccountDAO();
 
                 // Retrieve counts
-                List<Account> Accounts = udao.getAllAccWithUser();
+                udao.AccDelete(namedele);
 
-                // Retrieve booking details
-                request.setAttribute("accountusers", Accounts);
-
-                
-//                     UserAccountDAO udao = new UserAccountDAO();
-//
-//                // Retrieve counts
-//                List<UserAccount> Accounts = udao.getAllUserAccount();
-//
-//                // Retrieve booking details
-//                request.setAttribute("accountusers", Accounts);
-//                
+//            
                 // Forward the request to the admin JSP page
                 request.getRequestDispatcher("/admin/accountManager.jsp").forward(request, response);
 
             } else {
                 response.sendRedirect("login");
             }
+            
+            
         } catch (ServletException | IOException e) {
             response.sendRedirect("404.jsp");
         }
 
     }
 
-    public static void main(String[] args) {
-  AccountDAO udao = new AccountDAO();
-
-                // Retrieve counts
- 
-        
-        
-        
-        
-        
-        
-        
-    }
-
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /**
+    /** 
      * Handles the HTTP <code>GET</code> method.
-     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -105,13 +76,12 @@ public class accountManagement extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
+    throws ServletException, IOException {
         processRequest(request, response);
-    }
+    } 
 
-    /**
+    /** 
      * Handles the HTTP <code>POST</code> method.
-     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -119,13 +89,44 @@ public class accountManagement extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        processRequest(request, response);
+    throws ServletException, IOException {
+          try (PrintWriter out = response.getWriter()) {
+        HttpSession session = request.getSession(); 
+        model.Account acc = (model.Account) session.getAttribute("acc");
+
+        if (acc != null && acc.getRole_id().equalsIgnoreCase("1")) {
+            // Retrieve form data
+            String username = request.getParameter("username");
+            String password = request.getParameter("password");
+            String phone = request.getParameter("phone");
+            String email = request.getParameter("email");
+            int role = Integer.parseInt(request.getParameter("role")); // Assuming role is stored as integer in DB
+            boolean isActive = Boolean.parseBoolean(request.getParameter("isActive"));
+            String fullname = request.getParameter("fullname");
+            Date dob = Date.valueOf(request.getParameter("dob")); // Conversion to Date
+            boolean gender = Boolean.parseBoolean(request.getParameter("gender"));
+            String address = request.getParameter("address");
+
+            // Create an AccountDAO instance and call the method to create an account
+            AccountDAO accountDAO = new AccountDAO();
+           
+            accountDAO.updateAccountAndUser(username, password, phone, email, role, isActive, fullname, dob, gender, address, gender);
+           
+
+            request.getRequestDispatcher("accountmanagement").forward(request, response);
+        } else {
+            response.sendRedirect("login");
+        }
+    } catch (ServletException | IOException e) {
+        response.sendRedirect("404.jsp");
+    }   catch (SQLException ex) {
+            Logger.getLogger(accEdit.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    
     }
 
-    /**
+    /** 
      * Returns a short description of the servlet.
-     *
      * @return a String containing servlet description
      */
     @Override
