+++ SWP291_Team2_SE1829NJ/src/java/controller/acc/accountManagement.java
@@ -2,7 +2,8 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-package controller.admin;
+
+package controller.acc;
 
 import dao.AccountDAO;
 import java.io.IOException;
@@ -19,7 +20,7 @@ import model.Account;
  *
  * @author HUNG
  */
-public class accDelete extends HttpServlet {
+public class accountManagement extends HttpServlet {
 
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
@@ -32,8 +33,57 @@ public class accDelete extends HttpServlet {
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
+  
+        response.setCharacterEncoding("UTF-8");
+        response.setContentType("text/html; charset=UTF-8");
 
+        try {
+            HttpSession session = request.getSession(); 
+            model.Account acc = (Account) session.getAttribute("acc");
+            if (acc.getRole_id().equalsIgnoreCase("1")) {
+
+//             
+                AccountDAO udao = new AccountDAO();
+
+                // Retrieve counts
+                List<Account> Accounts = udao.getAllAccWithUser();
+
+                // Retrieve booking details
+                request.setAttribute("accountusers", Accounts);
+
+                
+//                     UserAccountDAO udao = new UserAccountDAO();
+//
+//                // Retrieve counts
+//                List<UserAccount> Accounts = udao.getAllUserAccount();
+//
+//                // Retrieve booking details
+//                request.setAttribute("accountusers", Accounts);
+//                
+                // Forward the request to the admin JSP page
+                request.getRequestDispatcher("/admin/accountManager.jsp").forward(request, response);
+
+            } else {
+                response.sendRedirect("login");
+            }
+        } catch (ServletException | IOException e) {
+            response.sendRedirect("404.jsp");
+        }
+
+    }
+
+    public static void main(String[] args) {
+  AccountDAO udao = new AccountDAO();
+
+                // Retrieve counts
+ 
+        
+        
+        
+        
+        
+        
+        
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
@@ -62,32 +112,7 @@ public class accDelete extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setCharacterEncoding("UTF-8");
-        response.setContentType("text/html; charset=UTF-8");
-
-        try {
-            HttpSession session = request.getSession();
-            model.Account acc = (Account) session.getAttribute("acc");
-            if (acc.getRole_id().equalsIgnoreCase("1")) {
-
-//             
-                String namedele = request.getParameter("username");
-                System.out.println("Username to delete: " + namedele); // Kiểm tra giá trị nhận được
-                if (namedele != null) {
-                    AccountDAO udao = new AccountDAO();
-                    udao.AccDelete(namedele);
-                    request.getRequestDispatcher("accountmanagement").forward(request, response);
-                } else {
-                    System.out.println("No username received for deletion.");
-                    response.sendRedirect("errorPage.jsp"); // Chuyển hướng đến trang lỗi nếu không có dữ liệu
-                }
-
-            } else {
-                response.sendRedirect("login");
-            }
-        } catch (ServletException | IOException e) {
-            response.sendRedirect("404.jsp");
-        }
+        processRequest(request, response);
     }
 
     /**
