+++ SWP291_Team2_SE1829NJ/web/WEB-INF/web.xml
@@ -8,12 +8,10 @@
         <servlet-name>RegisterController</servlet-name>
         <servlet-class>controller.RegisterController</servlet-class>
     </servlet>
- 
     <servlet>
         <servlet-name>roomManagement</servlet-name>
         <servlet-class>controller.admin.roomManagement</servlet-class>
     </servlet>
-
     <servlet>
         <servlet-name>ListManageBlogServlet</servlet-name>
         <servlet-class>controller.marketer.ListManageBlogServlet</servlet-class>
@@ -38,7 +36,6 @@
         <servlet-name>homepage</servlet-name>
         <servlet-class>controller.homepage</servlet-class>
     </servlet>
-
     <servlet>
         <servlet-name>roomManage</servlet-name>
         <servlet-class>controller.manager.roomManage</servlet-class>
@@ -55,6 +52,26 @@
         <servlet-name>roomEdit</servlet-name>
         <servlet-class>controller.manager.roomEdit</servlet-class>
     </servlet>
+    <servlet>
+        <servlet-name>Dashboard</servlet-name>
+        <servlet-class>controller.dashboard.Dashboard</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>accCreate</servlet-name>
+        <servlet-class>controller.acc.accCreate</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>accDelete</servlet-name>
+        <servlet-class>controller.acc.accDelete</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>accEdit</servlet-name>
+        <servlet-class>controller.acc.accEdit</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>accountManagement</servlet-name>
+        <servlet-class>controller.acc.accountManagement</servlet-class>
+    </servlet>
     <servlet-mapping>
         <servlet-name>LoginController</servlet-name>
         <url-pattern>/login</url-pattern>
@@ -63,8 +80,6 @@
         <servlet-name>RegisterController</servlet-name>
         <url-pattern>/register</url-pattern>
     </servlet-mapping>
- 
-
     <servlet-mapping>
         <servlet-name>ListManageBlogServlet</servlet-name>
         <url-pattern>/listmanageblog</url-pattern>
@@ -89,7 +104,6 @@
         <servlet-name>homepage</servlet-name>
         <url-pattern>/homepage</url-pattern>
     </servlet-mapping>
-
     <servlet-mapping>
         <servlet-name>roomManage</servlet-name>
         <url-pattern>/roommanagement</url-pattern>
@@ -105,5 +119,25 @@
     <servlet-mapping>
         <servlet-name>roomEdit</servlet-name>
         <url-pattern>/roomedit</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>Dashboard</servlet-name>
+        <url-pattern>/dashboard</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>accCreate</servlet-name>
+        <url-pattern>/acccreate</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>accDelete</servlet-name>
+        <url-pattern>/accdelete</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>accEdit</servlet-name>
+        <url-pattern>/accedit</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>accountManagement</servlet-name>
+        <url-pattern>/accountmanagement</url-pattern>
     </servlet-mapping>
 </web-app>
